--- kratos/python_scripts/run_cpp_tests.py
@@ -1,9 +1,0 @@
-import KratosMultiphysics as Kratos
-import sys
-if len(sys.argv) < 2:
-    Kratos.Tester.SetVerbosity(Kratos.Tester.Verbosity.TESTS_LIST)
-    Kratos.Tester.RunAllTestCases()
-else :
-    Kratos.Tester.SetVerbosity(Kratos.Tester.Verbosity.TESTS_OUTPUTS)
-    Kratos.Tester.RunTestCases(sys.argv[1])
-
