+++ kratos/python_scripts/testing/run_cpp_mpi_tests.py
@@ -1,0 +1,7 @@
+from KratosMultiphysics import *
+from KratosMultiphysics import mpi
+
+#Tester.SetVerbosity(Tester.Verbosity.TESTS_LIST)
+Tester.SetVerbosity(Tester.Verbosity.TESTS_OUTPUTS)
+
+Tester.RunTestSuite("KratosMPICoreFastSuite")
