+++ kratos/python_scripts/testing/run_cpp_tests.py
@@ -1,0 +1,9 @@
+import KratosMultiphysics as Kratos
+import sys
+if len(sys.argv) < 2:
+    Kratos.Tester.SetVerbosity(Kratos.Tester.Verbosity.TESTS_LIST)
+    Kratos.Tester.RunAllTestCases()
+else :
+    Kratos.Tester.SetVerbosity(Kratos.Tester.Verbosity.TESTS_OUTPUTS)
+    Kratos.Tester.RunTestCases(sys.argv[1])
+
